wip: clean up tests
